Remove Marc's opinions about YAML.
Documentation is not the place for perosnal grudges and protocol slap
fights.  Nor is it the place to document every minute problem with the
YAML spec.  It's trolling with module documentation.  None of this
helps the module user.

The code example to create valid YAML is useful, I left that in.  If
there is great concern about producing YAML compatible output, an
as_yaml method would be more useful to the user than enumerating all
the minute ways YAML is wrong.

YAML 1.2 went to great lengths to make YAML a true superset of JSON.
Most of the claims made here are out of date (Unicode encodings) or so
obscure I can't figure out if they apply (1024 char key limit).