Preserve types of numeric values better
fixed encode of numbers for dual-vars. Different string representations
are preserved, but numbers with temporary strings which represent the same
number are here treated as numbers, not strings.
Cpanel::JSON::XS is now a bit slower, but preserves numeric types better.
This is now the same as the PR #10 for JSON::PP by kraih.
https://github.com/makamaka/JSON-PP/pull/10/