Detect INF/NAN: ?/++/-?/--- on HP-UX (#56)
Add a seperate neg_nan, which is -? on HP-UX, but with our
tests nan is always -? and neg_nan is ?. Similar to Windows.

Note that detecting the sprintf value after the conversion is problematic.
The Apache stdcxx switched to use the system macros is_nan/is_inf/... before.
HP-UX has the known libc bug not supporting negative NAN,
https://issues.apache.org/jira/browse/STDCXX-464
HP-UX may return ?.00000