Check d_isinf/d_isnan before, no sprintf
It is better to check the system provided isinf/isnan
before, than to rely on the sprintf platform quirks.
This should also handle the unhandled OSF/1, and later
fixes in various libc's (Solaris, HP-UX, newlib, mingw).

But we can use this only for mode 0, which does not require
returning the platform-specific stringified value.
And for mode 3, we have no API to check for -inf.