add BOM support to decode (WIP)
Detect BOM and convert to UTF-8 and set UTF8 flag.

https://tools.ietf.org/html/rfc7159#section-8.1
JSON text SHALL be encoded in UTF-8, UTF-16, or UTF-32.

Byte Order Mark - While section 8.1 states "Implementations MUST
NOT add a byte order mark to the beginning of a JSON text",
"implementations (...) MAY ignore the presence of a byte order
mark rather than treating it as an error".