BOM: handle multibyte BOM, fix some offset and double-encoding
to enable a proper free of the old PV.
Note that decode_bom does it internally, and it returns a fresh SV.

Encode::decodes creates a copy,
Encode::from_to converts in-place.
Let decode detect the BOM and skip it internally. which is much
easier than with from_to() in-place. decode_xs() would be even faster.

TODO: A wrong refcount somewhere. Don't encode multibyte to UTF-8 yet.