Fix typed json encoder conversion from scalar's PV slot to JSON_TYPE_INT
It did not worked correctly for numbers which can be represented by
unsigned type but not by signed type: [2^(8*IVSIZE-1), 2^(8*IVSIZE)-1]

SvIV and SvUV macros may modify SvIsUV flag, so after calling SvIV it is
needed to check if returned value is signed or unsigned.

This patch fixes it and add tests for scalars with corner cases in PV slot.

Values out of range of perl integers are converted for JSON_TYPE_INT to
either IV_MIN or UV_MAX (signed minimal or unsigned maximal represented
value in perl).