For JSON_TYPE_INT encoder allow to pass Math::BigInt and Math::BigFloat objects
Option ->allow_bignum is required. Without it it is not possible to pass
Math::BigInt and Math::BigFloat objects.

For converting Math::BigFloat to Math::BigInt should be used $obj->bint
method, equivalent of perl's int() function for non-big floats. But this
method is not available in older Math::BigFloat versions. Older
Math::BigFloat versions have $obj->as_number method which should do same
thing as $obj->bint method but in some Math::BigFloat versions it is broken
and loose precision, so it cannot be used.

So for conversion from Math::BigFloat to Math::BigInt use either
$obj->bceil or $obj->bfloor method based on information if $obj is negative
or not. This is simulation of rounding and should give us same result as
$obj->bint method.