fix canonical with tied hashes (GH #167)
This feature was deliberately skipped in JSON::XS,
because it is a bit convoluted to sort tied hashes.

Create fresh HEs because the next iter will free it.

Also optimize sorting of hashes of length 1.

Also optimize the handling of utf8 keys. Only when one such key
is detected sort with use bytes in scope, which is a bit slower.
Fix use bytes handling for >=5.10, we need to set cop_hints,
not op_private.