Fix decode_json(string, 0)
check for the actual true-ness of the optional 2nd decode_json argument,
as described in the docs. So far only the existness of an 2nd argument
rendered it true, but with the 3rd optional arg it became an even worse bug.
SvTRUE did exist in 5.6 already.

Fixes GH #171

Also improve the docs around it.