remove overloads before re-adding them to silence warnings
To prevent warnings from redefining overloads on JSON::PP::Boolean, the
code is setting $^W to 0. This has classically worked because
overload.pm did not enable or disable warnings. This is not something
that should be relied on.

Rather than rely on $^W, just remove the overloads we are expecting to
overwrite using overload->unimport.